12-ml_enable_agent/requirements.txt
# Iris Classification Enable Agent Dependencies

# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# OpenAI
openai>=1.12.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
